Switch more small values to shorts or unsigned shorts...

+++ kfmon.c
@@ -239,13 +239,13 @@ static void
 	int           mfd = open("/proc/mounts", O_RDONLY, 0);
 	struct pollfd pfd;
 
-	unsigned int changes = 0;
-	pfd.fd               = mfd;
-	pfd.events           = POLLERR | POLLPRI;
-	pfd.revents          = 0;
+	unsigned short int changes = 0;
+	pfd.fd                     = mfd;
+	pfd.events                 = POLLERR | POLLPRI;
+	pfd.revents                = 0;
 	while (poll(&pfd, 1, -1) >= 0) {
 		if (pfd.revents & POLLERR) {
-			LOG(LOG_INFO, "Mountpoints changed (iteration nr. %u)", changes++);
+			LOG(LOG_INFO, "Mountpoints changed (iteration nr. %hu)", changes++);
 
 			// Stop polling once we know our mountpoint is available...
 			if (is_target_mounted()) {
@@ -408,8 +408,8 @@ static bool
 	} else {
 		// Make sure we're not trying to set multiple watches on the same file...
 		// (because that would only actually register the first one parsed).
-		unsigned int watch_idx = 0;
-		unsigned int matches   = 0;
+		unsigned short int watch_idx = 0;
+		unsigned short int matches   = 0;
 		for (watch_idx = 0; watch_idx < WATCH_MAX; watch_idx++) {
 			if (strcmp(pconfig->filename, watch_config[watch_idx].filename) == 0) {
 				matches++;
@@ -509,7 +509,7 @@ static int
 							rval = -1;
 						} else {
 							LOG(LOG_NOTICE,
-							    "Daemon config loaded from '%s': db_timeout=%u, use_syslog=%u, with_notifications=%u",
+							    "Daemon config loaded from '%s': db_timeout=%hu, use_syslog=%hu, with_notifications=%hu",
 							    p->fts_name,
 							    daemon_config.db_timeout,
 							    daemon_config.use_syslog,
@@ -538,7 +538,7 @@ static int
 						} else {
 							if (validate_watch_config(&watch_config[watch_count])) {
 								LOG(LOG_NOTICE,
-								    "Watch config @ index %u loaded from '%s': filename=%s, action=%s, block_spawns=%u, do_db_update=%u, db_title=%s, db_author=%s, db_comment=%s",
+								    "Watch config @ index %hu loaded from '%s': filename=%s, action=%s, block_spawns=%hu, do_db_update=%hu, db_title=%s, db_author=%s, db_comment=%s",
 								    watch_count,
 								    p->fts_name,
 								    watch_config[watch_count].filename,
@@ -573,13 +573,13 @@ static int
 
 #ifdef DEBUG
 	// Let's recap (including failures)...
-	DBGLOG("Daemon config recap: db_timeout=%u, use_syslog=%u, with_notifications=%u",
+	DBGLOG("Daemon config recap: db_timeout=%hu, use_syslog=%hu, with_notifications=%hu",
 	       daemon_config.db_timeout,
 	       daemon_config.use_syslog,
 	       daemon_config.with_notifications);
-	for (unsigned int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
+	for (unsigned short int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
 		DBGLOG(
-		    "Watch config @ index %u recap: filename=%s, action=%s, block_spawns=%u, skip_db_checks=%u, do_db_update=%u, db_title=%s, db_author=%s, db_comment=%s",
+		    "Watch config @ index %hu recap: filename=%s, action=%s, block_spawns=%hu, skip_db_checks=%hu, do_db_update=%hu, db_title=%s, db_author=%s, db_comment=%s",
 		    watch_idx,
 		    watch_config[watch_idx].filename,
 		    watch_config[watch_idx].action,
@@ -614,7 +614,7 @@ static unsigned int
 
 // Check if our target file has been processed by Nickel...
 static bool
-    is_target_processed(unsigned int watch_idx, bool wait_for_db)
+    is_target_processed(unsigned short int watch_idx, bool wait_for_db)
 {
 	sqlite3*      db;
 	sqlite3_stmt* stmt;
@@ -704,7 +704,7 @@ static bool
 			DBGLOG("Checking for thumbnails in '%s' . . .", images_path);
 
 			// Count the number of processed thumbnails we find...
-			unsigned int thumbnails_num = 0;
+			unsigned short int thumbnails_num = 0;
 
 			// Start with the full-size screensaver...
 			char ss_path[KFMON_PATH_MAX];
@@ -812,10 +812,10 @@ static bool
 		// NOTE: This assumes the DB was opened with the default journal_mode, DELETE
 		//       This doesn't appear to be the case anymore, on FW >= 4.6.x (and possibly earlier),
 		//       it's now using WAL (which makes sense).
-		unsigned int count = 0;
+		unsigned short int count = 0;
 		while (access(KOBO_DB_PATH "-journal", F_OK) == 0) {
 			LOG(LOG_INFO,
-			    "Found a SQLite rollback journal, waiting for it to go away (iteration nr. %u) . . .",
+			    "Found a SQLite rollback journal, waiting for it to go away (iteration nr. %hu) . . .",
 			    count++);
 			nanosleep((const struct timespec[]){ { 0, 500000000L } }, NULL);
 			// NOTE: Don't wait more than 10s
@@ -837,17 +837,17 @@ static bool
 static void
     init_process_table(void)
 {
-	for (unsigned int i = 0; i < WATCH_MAX; i++) {
+	for (unsigned short int i = 0; i < WATCH_MAX; i++) {
 		PT.spawn_pids[i]     = -1;
 		PT.spawn_watchids[i] = -1;
 	}
 }
 
 // Returns the index of the next available entry in the process table.
-static int
+static short int
     get_next_available_pt_entry(void)
 {
-	for (int i = 0; i < WATCH_MAX; i++) {
+	for (short int i = 0; i < WATCH_MAX; i++) {
 		if (PT.spawn_watchids[i] == -1) {
 			return i;
 		}
@@ -857,15 +857,15 @@ static int
 
 // Adds information about a new spawn to the process table.
 static void
-    add_process_to_table(int i, pid_t pid, unsigned int watch_idx)
+    add_process_to_table(short int i, pid_t pid, unsigned short int watch_idx)
 {
 	PT.spawn_pids[i]     = pid;
-	PT.spawn_watchids[i] = (int) watch_idx;
+	PT.spawn_watchids[i] = (short int) watch_idx;
 }
 
 // Removes information about a spawn from the process table.
 static void
-    remove_process_from_table(int i)
+    remove_process_from_table(short int i)
 {
 	PT.spawn_pids[i]     = -1;
 	PT.spawn_watchids[i] = -1;
@@ -893,13 +893,13 @@ static void
 void*
     reaper_thread(void* ptr)
 {
-	int i = *((int*) ptr);
+	short int i = *((short int*) ptr);
 
 	pid_t tid;
 	tid = (pid_t) syscall(SYS_gettid);
 
-	pid_t cpid;
-	int   watch_idx;
+	pid_t     cpid;
+	short int watch_idx;
 	pthread_mutex_lock(&ptlock);
 	cpid      = PT.spawn_pids[i];
 	watch_idx = PT.spawn_watchids[i];
@@ -912,7 +912,7 @@ void*
 	// Remember the current time for the execvp errno/exitcode heuristic...
 	time_t then = time(NULL);
 
-	MTLOG("[%s] [INFO] [TID: %ld] Waiting to reap process %ld (from watch idx %d) . . .",
+	MTLOG("[%s] [INFO] [TID: %ld] Waiting to reap process %ld (from watch idx %hd) . . .",
 	      get_current_time_r(&local_tm, sz_time, sizeof(sz_time)),
 	      (long) tid,
 	      (long) cpid,
@@ -932,7 +932,7 @@ void*
 		if (WIFEXITED(wstatus)) {
 			int exitcode = WEXITSTATUS(wstatus);
 			MTLOG(
-			    "[%s] [NOTE] [TID: %ld] Reaped process %ld (from watch idx %d): It exited with status %d.",
+			    "[%s] [NOTE] [TID: %ld] Reaped process %ld (from watch idx %hd): It exited with status %d.",
 			    get_current_time_r(&local_tm, sz_time, sizeof(sz_time)),
 			    (long) tid,
 			    (long) cpid,
@@ -968,7 +968,7 @@ void*
 			snprintf(
 			    buf,
 			    sizeof(buf),
-			    "[KFMon] [%s] [WARN] [TID: %ld] Reaped process %ld (from watch idx %d): It was killed by signal %d",
+			    "[KFMon] [%s] [WARN] [TID: %ld] Reaped process %ld (from watch idx %hd): It was killed by signal %d",
 			    get_current_time_r(&local_tm, sz_time, sizeof(sz_time)),
 			    (long) tid,
 			    (long) cpid,
@@ -1001,7 +1001,7 @@ void*
 // As well as the glibc's system() call,
 // With a bit of added tracking to handle reaping without a SIGCHLD handler.
 static pid_t
-    spawn(char* const* command, unsigned int watch_idx)
+    spawn(char* const* command, unsigned short int watch_idx)
 {
 	pid_t pid;
 
@@ -1038,7 +1038,7 @@ static pid_t
 	} else {
 		// Parent
 		// Keep track of the process
-		int i;
+		short int i;
 		pthread_mutex_lock(&ptlock);
 		i = get_next_available_pt_entry();
 		pthread_mutex_unlock(&ptlock);
@@ -1059,14 +1059,14 @@ static pid_t
 			add_process_to_table(i, pid, watch_idx);
 			pthread_mutex_unlock(&ptlock);
 
-			DBGLOG("Assigned pid %ld (from watch idx %u) to process table entry idx %d",
+			DBGLOG("Assigned pid %ld (from watch idx %hu) to process table entry idx %hd",
 			       (long) pid,
 			       watch_idx,
 			       i);
 			// NOTE: We can't do that from the child proper, because it's not async-safe,
 			//       so do it from here.
 			LOG(LOG_NOTICE,
-			    "Spawned process %ld (%s -> %s @ watch idx %u) . . .",
+			    "Spawned process %ld (%s -> %s @ watch idx %hu) . . .",
 			    (long) pid,
 			    watch_config[watch_idx].filename,
 			    watch_config[watch_idx].action,
@@ -1080,8 +1080,8 @@ static pid_t
 			// NOTE: We achieve reaping in a non-blocking way by doing the reaping from a dedicated thread
 			//       for every spawn...
 			//       See #2 for an history of the previous failed attempts...
-			pthread_t rthread;
-			int*      arg = malloc(sizeof(*arg));
+			pthread_t  rthread;
+			short int* arg = malloc(sizeof(*arg));
 			if (arg == NULL) {
 				LOG(LOG_ERR, "Couldn't allocate memory for thread arg, aborting!");
 				fbink_print(-1, "[KFMon] OOM ?!", &fbink_config);
@@ -1142,11 +1142,11 @@ static pid_t
 
 // Check if a given inotify watch already has a spawn running
 static bool
-    is_watch_already_spawned(unsigned int watch_idx)
+    is_watch_already_spawned(unsigned short int watch_idx)
 {
 	// Walk our process table to see if the given watch currently has a registered running process
-	for (unsigned int i = 0; i < WATCH_MAX; i++) {
-		if (PT.spawn_watchids[i] == (int) watch_idx) {
+	for (unsigned short int i = 0; i < WATCH_MAX; i++) {
+		if (PT.spawn_watchids[i] == (short int) watch_idx) {
 			return true;
 			// NOTE: Assume everything's peachy,
 			//       and we'll never end up with the same watch_idx assigned to multiple indices in the
@@ -1165,11 +1165,11 @@ static bool
     is_blocker_running(void)
 {
 	// Walk our process table to identify watches with a currently running process
-	for (unsigned int i = 0; i < WATCH_MAX; i++) {
+	for (unsigned short int i = 0; i < WATCH_MAX; i++) {
 		if (PT.spawn_watchids[i] != -1) {
 			// Walk the registered watch list to match that currently running watch to its block_spawns flag
-			for (unsigned int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
-				if (PT.spawn_watchids[i] == (int) watch_idx) {
+			for (unsigned short int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
+				if (PT.spawn_watchids[i] == (short int) watch_idx) {
 					if (watch_config[watch_idx].block_spawns) {
 						return true;
 					}
@@ -1184,10 +1184,10 @@ static bool
 
 // Return the pid of the spawn of a given inotify watch
 static pid_t
-    get_spawn_pid_for_watch(unsigned int watch_idx)
+    get_spawn_pid_for_watch(unsigned short int watch_idx)
 {
-	for (unsigned int i = 0; i < WATCH_MAX; i++) {
-		if (PT.spawn_watchids[i] == (int) watch_idx) {
+	for (unsigned short int i = 0; i < WATCH_MAX; i++) {
+		if (PT.spawn_watchids[i] == (short int) watch_idx) {
 			return PT.spawn_pids[i];
 		}
 	}
@@ -1236,8 +1236,8 @@ static bool
 			// memcpy(&event, &ptr, sizeof(struct inotify_event *));
 
 			// Identify which of our target file we've caught an event for...
-			unsigned int watch_idx       = 0;
-			bool         found_watch_idx = false;
+			unsigned short int watch_idx       = 0;
+			bool               found_watch_idx = false;
 			for (watch_idx = 0; watch_idx < watch_count; watch_idx++) {
 				if (watch_config[watch_idx].inotify_wd == event->wd) {
 					found_watch_idx = true;
@@ -1318,7 +1318,7 @@ static bool
 					// before launching anything...
 					if (!pending_processing && is_target_processed(watch_idx, true)) {
 						LOG(LOG_INFO,
-						    "Preparing to spawn %s for watch idx %u . . .",
+						    "Preparing to spawn %s for watch idx %hu . . .",
 						    watch_config[watch_idx].action,
 						    watch_idx);
 						if (watch_config[watch_idx].block_spawns) {
@@ -1348,7 +1348,7 @@ static bool
 						pthread_mutex_unlock(&ptlock);
 
 						LOG(LOG_INFO,
-						    "As watch idx %u (%s) still has a spawned process (%ld -> %s) running, we won't be spawning another instance of it!",
+						    "As watch idx %hu (%s) still has a spawned process (%ld -> %s) running, we won't be spawning another instance of it!",
 						    watch_idx,
 						    watch_config[watch_idx].filename,
 						    (long) spid,
@@ -1396,7 +1396,7 @@ static bool
 				// Try to remove the inotify watch we matched
 				// (... hoping matching actually was successful), and break the loop.
 				LOG(LOG_INFO,
-				    "Trying to remove inotify watch for '%s' @ index %u.",
+				    "Trying to remove inotify watch for '%s' @ index %hu.",
 				    watch_config[watch_idx].filename,
 				    watch_idx);
 				if (inotify_rm_watch(fd, watch_config[watch_idx].inotify_wd) == -1) {
@@ -1416,7 +1416,7 @@ static bool
 		if (destroyed_wd) {
 			// But before we do that, make sure we've removed *all* our *other* watches first
 			// (again, hoping matching was successful), since we'll be setting them up all again later...
-			for (unsigned int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
+			for (unsigned short int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
 				if (!watch_config[watch_idx].wd_was_destroyed) {
 					// Don't do anything if that was because of an unmount...
 					// Because that assures us that everything is/will soon be gone
@@ -1426,7 +1426,7 @@ static bool
 					if (!was_unmounted) {
 						// Log what we're doing...
 						LOG(LOG_INFO,
-						    "Trying to remove inotify watch for '%s' @ index %u.",
+						    "Trying to remove inotify watch for '%s' @ index %hu.",
 						    watch_config[watch_idx].filename,
 						    watch_idx);
 						if (inotify_rm_watch(fd, watch_config[watch_idx].inotify_wd) == -1) {
@@ -1573,7 +1573,7 @@ int
 		//       Relative to the earlier IN_MOVE_SELF mention, that means it'll keep tracking the file with its
 		//           new name (provided it was moved to the *same* fs,
 		//           as crossing a fs boundary will delete the original).
-		for (unsigned int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
+		for (unsigned short int watch_idx = 0; watch_idx < watch_count; watch_idx++) {
 			watch_config[watch_idx].inotify_wd =
 			    inotify_add_watch(fd, watch_config[watch_idx].filename, IN_OPEN | IN_CLOSE);
 			if (watch_config[watch_idx].inotify_wd == -1) {
@@ -1590,7 +1590,7 @@ int
 				//       during Nickel's processing of said target file ;).
 			}
 			LOG(LOG_NOTICE,
-			    "Setup an inotify watch for '%s' @ index %u.",
+			    "Setup an inotify watch for '%s' @ index %hu.",
 			    watch_config[watch_idx].filename,
 			    watch_idx);
 		}
